Implement cars' delete method

+++ pkg/handlers/cars.go
@@ -207,6 +207,15 @@ func (h *Handler) CarsUpdate(c *gin.Context) {
 
 // CarsDelete deletes a single existing manufacturer
 func (h *Handler) CarsDelete(c *gin.Context) {
+	manufacturerID, err := strconv.Atoi(c.Param("manufacturerID"))
+	if err != nil {
+		log.Println(err)
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Bad identifier",
+		})
+		return
+	}
+
 	carID, err := strconv.Atoi(c.Param("carID"))
 	if err != nil {
 		log.Println(err)
@@ -216,9 +225,9 @@ func (h *Handler) CarsDelete(c *gin.Context) {
 		return
 	}
 
-	var car models.Car
+	var manufacturer models.Manufacturer
 
-	result := h.DB.First(&car, carID)
+	result := h.DB.First(&manufacturer, manufacturerID)
 	if result.Error != nil {
 		log.Println(result.Error)
 		c.JSON(http.StatusNotFound, gin.H{
@@ -227,10 +236,28 @@ func (h *Handler) CarsDelete(c *gin.Context) {
 		return
 	}
 
-	h.DB.Delete(&car)
+	var car models.Car
+
+	// todo: look for another ways to check not found record error
+	result = h.DB.Where("manufacturer_id = ?", manufacturerID).Find(&car, carID)
+	if result.Error != nil || result.RowsAffected == 0 {
+		log.Println("Record not found", result.Error)
+		c.JSON(http.StatusNotFound, gin.H{
+			"message": "Record not found",
+		})
+		return
+	}
+
+	result = h.DB.Delete(&car)
+	if result.Error != nil {
+		log.Println(result.Error)
+		c.JSON(http.StatusInternalServerError, gin.H{
+			"message": "Internal server error",
+		})
+		return
+	}
 
 	c.JSON(http.StatusOK, gin.H{
 		"message": "Deleted car " + strconv.Itoa(carID),
 	})
-
 }
